requirements.txt
pinax-theme-bootstrap==7.1.1
metron==1.3.5
pinax-eventlog==1.0.0
django-jsonfield==0.9.15
dj-static==0.0.6
dj-database-url==0.3.0
gunicorn==19.3.0
psycopg2==2.6.1
pinax-pages==0.3.1
pinax-boxes==2.1.0
taggit-selectize==0.1.2
unicodecsv==0.14.1
mdx_linkify==0.6

# Those two packages are already installed by symposion/requirements.txt
#Django==1.8.5
#django-user-accounts==1.2
